requirements.txt
sentence-transformers>=2.2.2
faiss-cpu
rank-bm25
numpy
tqdm
fastapi
uvicorn
psutil
